pyproject.toml
[tool.poetry]
name = "fed-liquidity-monitor"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.46"
matplotlib = "^3.10.8"
pandas = "^3.0.0"
python-dotenv = "^1.2.1"
requests = "^2.32.5"
pyyaml = "^6.0.3"
python-dateutil = "^2.9.0.post0"
boto3 = "^1.42.50"
uvicorn = "^0.41.0"
fastapi = "^0.129.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
